Brought back logical_system_integ_test.go since I forgot we can't put those tests inside the "real" vault package due to import cycles.
